KREST-2654 Rate lmit per clusterId
Before this change, in a multi-tennant cluster, the overall rate limit for produce would be contributed to by all all the sepaate logical clusters.

This meant that one customer could use up the whole quote for the PKC for their single LKC, effectively blocking other customers from using their clusters.

This commit updates the rate limit to be per cluster id, so that each LKC has its own rate limit.

We'll have to carefully consider what to set the limit to, as if there are 1000 LKCs we need to be careful not to swamp the cluster.